requirements.txt
discord.py
wavelink
spotipy
requests
Pillow